// typescript-projects/utility-backups/src20250528BeforeEbillsTokenFromEnv/interface/types.ts
export type Flatten<T> = T extends Record<string, unknown>
  ? UnionToIntersection<
      {
        [K in keyof DeepRequired<T>]: K extends string
          ? T[K] extends Record<string, unknown>
            ? {
                [P in `${K}.${keyof Flatten<T[K]> & string}`]: Flatten<
                  T[K]
                >[P extends `${K}.${infer R}` ? R : never];
              } & { [P in K]: T[K] }
            : { [P in K]: T[K] }
          : never;
      }[keyof T]
    >
  : never;

export type UnionToIntersection<T> = (
  T extends unknown ? (x: T) => void : never
) extends (x: infer R) => void
  ? R
  : never;

/** Recursively make all properties of a map/object required */
export type DeepRequired<T> = {
  [K in keyof T]-?: Exclude<T[K], undefined> extends Record<string, unknown>
    ? DeepRequired<T[K]>
    : T[K];
};
// ...
export type TStatusDocsAccessData = {
  uid: string;
  /** hash of naira withdrawal tx, which also serves as tx doc ID */
  appTxHash: string;
  /** full path to the related naira_admin doc */
  nairaAdminDocPath: string;
};
// ...
export type TUtilityWithdrawalWriteData = {
  utility: {
    /** Wallet balance (at specific provider) before payment request is made */
    balanceBefore?: number;
    /** Wallet balance (at specific provider) after payment request is made */
    balanceAfter?: number;
    status?: AppTxStatus;
    /** ID of the utility payment tx we generate and supply to the provider */
    appOrderId?: string;
    /** ID of the utility payment tx the provider generates and sends back to us */
    providerOrderId?: string;
    /** The time at which the request was received and initiated by our app */
    initiatedAt?: number;
    /** The time at which a response is recieved from the provider */
    acknowledgedAt?: number;
    /** The time at which the utility tx gets a settled status like `approved` and `rejected` */
    settledAt?: number;
    providerId?: ProviderId;
    /** Electricity token */
    token?: string;
  };
};

export type TNairaAdminTxWriteData = {
  utilityStatus?: AppTxStatus;
};